Complete API for action cancelation.
